Move contexts and results to top-level API
