+++ subprojects/core/src/test/groovy/org/gradle/api/internal/tasks/execution/ExecuteActionsTaskExecuterTest.groovy
@@ -33,19 +33,19 @@ import org.gradle.groovy.scripts.ScriptSource
 import org.gradle.initialization.DefaultBuildCancellationToken
 import org.gradle.internal.exceptions.DefaultMultiCauseException
 import org.gradle.internal.exceptions.MultiCauseException
+import org.gradle.internal.execution.IncrementalChangesContext
+import org.gradle.internal.execution.IncrementalContext
 import org.gradle.internal.execution.OutputChangeListener
+import org.gradle.internal.execution.UpToDateResult
 import org.gradle.internal.execution.history.ExecutionHistoryStore
 import org.gradle.internal.execution.history.OutputFilesRepository
 import org.gradle.internal.execution.impl.DefaultWorkExecutor
-import org.gradle.internal.execution.impl.steps.CancelExecutionStep
-import org.gradle.internal.execution.impl.steps.CatchExceptionStep
-import org.gradle.internal.execution.impl.steps.ExecuteStep
-import org.gradle.internal.execution.impl.steps.IncrementalChangesContext
-import org.gradle.internal.execution.impl.steps.IncrementalContext
-import org.gradle.internal.execution.impl.steps.ResolveChangesStep
-import org.gradle.internal.execution.impl.steps.SkipUpToDateStep
-import org.gradle.internal.execution.impl.steps.SnapshotOutputStep
-import org.gradle.internal.execution.impl.steps.UpToDateResult
+import org.gradle.internal.execution.steps.CancelExecutionStep
+import org.gradle.internal.execution.steps.CatchExceptionStep
+import org.gradle.internal.execution.steps.ExecuteStep
+import org.gradle.internal.execution.steps.ResolveChangesStep
+import org.gradle.internal.execution.steps.SkipUpToDateStep
+import org.gradle.internal.execution.steps.SnapshotOutputStep
 import org.gradle.internal.id.UniqueId
 import org.gradle.internal.operations.BuildOperationContext
 import org.gradle.internal.operations.BuildOperationExecutor
