+++ subprojects/execution/src/main/java/org/gradle/internal/execution/CachingContext.java
@@ -14,19 +14,8 @@
  * limitations under the License.
  */
 
-package org.gradle.internal.execution.impl.steps;
+package org.gradle.internal.execution;
 
-import com.google.common.collect.ImmutableSortedMap;
-import org.gradle.caching.internal.origin.OriginMetadata;
-import org.gradle.internal.execution.Result;
-import org.gradle.internal.fingerprint.FileCollectionFingerprint;
-
-public interface SnapshotResult extends Result {
-    ImmutableSortedMap<String, ? extends FileCollectionFingerprint> getFinalOutputs();
-    OriginMetadata getOriginMetadata();
-
-    /**
-     * Did we reuse the output from some previous execution?
-     */
-    boolean isReused();
+public interface CachingContext extends IncrementalChangesContext {
+    CacheHandler getCacheHandler();
 }
