+++ subprojects/execution/src/main/java/org/gradle/internal/execution/UpToDateResult.java
@@ -14,12 +14,14 @@
  * limitations under the License.
  */
 
-package org.gradle.internal.execution.impl.steps;
+package org.gradle.internal.execution;
 
-import com.google.common.collect.ImmutableSortedMap;
-import org.gradle.internal.fingerprint.CurrentFileCollectionFingerprint;
+import com.google.common.collect.ImmutableList;
 
-public interface CurrentSnapshotResult extends SnapshotResult {
-    @Override
-    ImmutableSortedMap<String, CurrentFileCollectionFingerprint> getFinalOutputs();
+public interface UpToDateResult extends SnapshotResult {
+    /**
+     * A list of messages describing the first few reasons encountered that caused the work to be executed.
+     * An empty list means the work was up-to-date and hasn't been executed.
+     */
+    ImmutableList<String> getExecutionReasons();
 }
