+++ subprojects/execution/src/main/java/org/gradle/internal/execution/steps/TimeoutStep.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.gradle.internal.execution.impl.steps;
+package org.gradle.internal.execution.steps;
 
 import org.gradle.api.GradleException;
 import org.gradle.api.InvalidUserDataException;
